refactor(core): extract invocation pipeline from Crust
