+++ .changeset/spicy-suits-find.md
@@ -1,0 +1,5 @@
+---
+"@crustjs/core": patch
+---
+
+Refactor command invocation orchestration into a private pipeline module without changing behavior.
